Partial kettle rewrite

+++ common/src/main/java/net/favouriteless/enchanted/common/util/RandomUtils.java
@@ -31,7 +31,7 @@ public class RandomUtils {
     }
 
     public static float nextFloat() {
-        return RANDOM.nextInt();
+        return RANDOM.nextFloat();
     }
 
     public static float nextFloat(float upper) {
